+++ ipywidgets/widgets/widget_layout.py
@@ -24,6 +24,8 @@ class Layout(Widget):
     """
 
     _view_name = Unicode('LayoutView').tag(sync=True)
+    _view_module = Unicode('@jupyter-widgets/base')
+    _view_module_version = Unicode('3.0.0')
     _model_name = Unicode('LayoutModel').tag(sync=True)
 
     # Keys